refactor for new component structure
