+++ web/src/app/admin/embeddings/modals/AlreadyPickedModal.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Modal } from "@/components/Modal";
 import { Button, Text } from "@tremor/react";
 
-import { CloudEmbeddingModel } from "../components/types";
+import { CloudEmbeddingModel } from "../../../../components/embedding/types";
 
 export function AlreadyPickedModal({
   model,
