+++ web/src/components/embedding/EmbeddingSidebar.tsx
