+++ web/src/components/embedding/ModelSelector.tsx
