+++ web/src/components/embedding/ReindexingProgressTable.tsx
