+++ web/src/components/embedding/types.ts
